Falloff prevention added, wall now ‘solid’ & throwing algorithm started
Set wall sprite to ‘solid’ and started working on the throwing algorithm (this takes care of calculating the x and y translation of the bullet automatically based on the current angle). Sea Lion is now also prevented from falling off the screen when going too far to the left. BUG: when firing lots of fishes, the physics-bodies crash which makes the sprites hang in the air.